Feature/dev

+++ src/Sfa.Tl.Matching.Data/Repositories/GenericRepository.cs
@@ -73,6 +73,8 @@ namespace Sfa.Tl.Matching.Data.Repositories
 
         public virtual async Task UpdateMany(IList<T> entities)
         {
+            if (entities.Count == 0) return;
+
             _dbContext.UpdateRange(entities);
 
             try
@@ -90,6 +92,28 @@ namespace Sfa.Tl.Matching.Data.Repositories
         {
             foreach (var property in properties)
                 _dbContext.Entry(entity).Property(property).IsModified = true;
+
+            try
+            {
+                await _dbContext.SaveChangesAsync();
+            }
+            catch (DbUpdateException due)
+            {
+                _logger.LogError(due.Message, due.InnerException);
+                throw;
+            }
+        }
+
+        public virtual async Task UpdateManyWithSpecifedColumnsOnly(IList<T> entities,
+            params Expression<Func<T, object>>[] properties)
+        {
+            foreach (var entity in entities)
+            {
+                foreach (var property in properties)
+                {
+                    _dbContext.Entry(entity).Property(property).IsModified = true;
+                }
+            }
 
             try
             {
@@ -143,6 +167,8 @@ namespace Sfa.Tl.Matching.Data.Repositories
 
         public virtual async Task DeleteMany(IList<T> entities)
         {
+            if (entities.Count == 0) return;
+             
             _dbContext.RemoveRange(entities);
 
             try
@@ -158,26 +184,60 @@ namespace Sfa.Tl.Matching.Data.Repositories
 
         public IQueryable<T> GetMany(Expression<Func<T, bool>> predicate = null, params Expression<Func<T, object>>[] navigationPropertyPath)
         {
-            var queryable = GetQueryableWithIncludes(navigationPropertyPath);
+            var queryable = GetQueryableWithIncludes(predicate, null, true, navigationPropertyPath);
 
             return predicate != null ? queryable.Where(predicate) : queryable;
         }
 
         public async Task<T> GetSingleOrDefault(Expression<Func<T, bool>> predicate, params Expression<Func<T, object>>[] navigationPropertyPath)
         {
-            var queryable = GetQueryableWithIncludes(navigationPropertyPath);
+            var queryable = GetQueryableWithIncludes(predicate, null, true, navigationPropertyPath);
 
-            return await queryable.SingleOrDefaultAsync(predicate);
+            return await queryable.SingleOrDefaultAsync();
         }
 
-        private IQueryable<T> GetQueryableWithIncludes(Expression<Func<T, object>>[] navigationPropertyPath)
+        public async Task<T> GetFirstOrDefault(Expression<Func<T, bool>> predicate, params Expression<Func<T, object>>[] navigationPropertyPath)
+        {
+            var queryable = GetQueryableWithIncludes(predicate, null, true, navigationPropertyPath);
+
+            return await queryable.FirstOrDefaultAsync();
+        }
+
+        public async Task<TDto> GetSingleOrDefault<TDto>(Expression<Func<T, bool>> predicate, Expression<Func<T, TDto>> selector, Expression<Func<T, object>> orderBy, bool asendingorder = true, params Expression<Func<T, object>>[] navigationPropertyPath)
+        {
+            var queryable = GetQueryableWithIncludes(predicate, orderBy, asendingorder, navigationPropertyPath);
+
+            return await queryable.Select(selector).SingleOrDefaultAsync();
+
+        }
+
+        public async Task<TDto> GetFirstOrDefault<TDto>(Expression<Func<T, bool>> predicate, Expression<Func<T, TDto>> selector, Expression<Func<T, object>> orderBy, bool asendingorder = true, params Expression<Func<T, object>>[] navigationPropertyPath)
+        {
+            var queryable = GetQueryableWithIncludes(predicate, orderBy, asendingorder, navigationPropertyPath);
+
+            return await queryable.Select(selector).FirstOrDefaultAsync();
+        }
+
+        public async Task<int> Count(Expression<Func<T, bool>> predicate = null)
+        {
+            return predicate != null ? await _dbContext.Set<T>().CountAsync(predicate) :
+                await _dbContext.Set<T>().CountAsync();
+        }
+
+        private IQueryable<T> GetQueryableWithIncludes(Expression<Func<T, bool>> predicate, Expression<Func<T, object>> orderBy, bool asendingorder = true, params Expression<Func<T, object>>[] navigationPropertyPath)
         {
             var queryable = _dbContext.Set<T>().AsQueryable();
 
-            if (navigationPropertyPath.Any())
+            if (navigationPropertyPath != null && navigationPropertyPath.Any())
             {
                 queryable = navigationPropertyPath.Aggregate(queryable, (current, navProp) => current.Include(navProp));
             }
+
+            if (predicate != null)
+                queryable = queryable.Where(predicate);
+
+            if (orderBy != null)
+                queryable = asendingorder ? queryable.OrderBy(orderBy) : queryable.OrderByDescending(orderBy);
 
             return queryable;
         }
